--- i2c/methods.js
@@ -1,52 +1,0 @@
-const { channel } = require('diagnostics_channel');
-const i2c = require('i2c-bus');
-
-//
-function openI2C(channel = 1) {
-    return i2c.openSync(channel);
-}
-
-function closeI2C(dataChannel) {
-    dataChannel.closeSync();
-}
-
-function writeI2C(dataChannel, ...configs) {
-    return dataChannel.writeByteSync(...configs);
-}
-
-function readI2CRaw(dataChannel, ...configs) {
-    return dataChannel.readByteSync(...configs);
-}
-//
-
-
-//
-function readI2CAddress(dataChannel, channelAddress, dataAddress) {
-    const high = readI2CRaw(dataChannel, channelAddress, dataAddress)
-    const low = readI2CRaw(dataChannel, channelAddress, dataAddress + 1);
-    const value = (high << 8) | low;
-
-    if (value >= 0x8000) {
-        return -((65535 - value) + 1);
-    } else {
-        return value;
-    }
-}
-
-function readI2CData(dataChannel, channelAddress, dataMap) {
-    return Object.keys(dataMap).reduce((acc, key) => {
-        const address = Number('0x' + dataMap[key].toString(16));
-        acc[key] = readI2CAddress(dataChannel, channelAddress, address);
-
-        return acc;
-    }, {});
-}
-//
-
-
-module.exports = {
-    openI2C,
-    closeI2C,
-    readI2CData,
-    writeI2C
-}
